Prefer raw parsing over JSON in Apache and JavaRestCalls
When passing String as response type to REST call, it does not know how
to decode the result: either use it as String, like it is, or parse a
JSON string. Previously JSON was preferred in the Apache and Java
client but not in the Spring client. Modified the default RestParser
setting for the Apache and Java client to match the Spring client.