Add searchAfter option
We are using an updated document store in the background, now. The
manufacturer is discuraging the use of page index values and does not
support those values anymore. Alternatively, one should use the
searchAfter parameter, that was provided with the previous response.

We have adapter our client to this behavior, but as a result, the
pageIndex values provided by our ResultPage are not reliable anymore.

If you are using the "nextPage" method of the result, it will use
the searchAfter feature, now.